refactor: server commands names
