+++ .vscode/launch.json
@@ -9,6 +9,7 @@
             "program": "${workspaceFolder}",
             "args": [
                 "serve",
+                "authz-server"
             ],
         },
         {
@@ -18,6 +19,7 @@
             "mode": "auto",
             "program": "${workspaceFolder}",
             "args": [
+                "serve",
                 "sidecar-injector",
             ],
         }
