+++ charts/kyverno-sidecar-injector/values.yaml
@@ -211,6 +211,7 @@ containers:
 
     # -- Container args.
     args:
+      - serve
       - sidecar-injector
       - --address=:9443
       - --cert-file=/opt/kubernetes-sidecar-injector/certs/tls.crt
