+++ pkg/commands/root/command.go
@@ -1,7 +1,6 @@
 package root
 
 import (
-	"github.com/kyverno/kyverno-envoy-plugin/pkg/commands/inject"
 	"github.com/kyverno/kyverno-envoy-plugin/pkg/commands/serve"
 	"github.com/spf13/cobra"
 )
@@ -12,6 +11,5 @@ func Command() *cobra.Command {
 		Short: "kyverno-envoy-plugin is a plugin for Envoy",
 	}
 	root.AddCommand(serve.Command())
-	root.AddCommand(inject.Command())
 	return root
 }
